migrate address book to data dir
