+++ common/v2/database/data/index.ts
@@ -3,3 +3,4 @@ export * from './contracts';
 export { NETWORKS_CONFIG, NetworkConfig } from './networks';
 export { NODES_CONFIG } from './nodes';
 export { SCHEMA_BASE } from './schema';
+export { defaultAddressBook } from './addressBook';
