+++ common/v2/database/seed/index.ts
@@ -1,3 +1,3 @@
 export { devAccounts, DevAccount, SeedAssetBalance } from './accounts';
 export { assets as devAssets } from './assets';
-export { defaultAddressBook, addressBook as devContacts } from './addressBook';
+export { addressBook as devContacts } from './addressBook';
